modularized code and settings
