+++ lib/globals.go
@@ -1,0 +1,11 @@
+package lib
+
+/*
+   File where only global constants are stored
+
+   v1.0 @AOrps
+*/
+
+const (
+	TEMPLATEPATH = "templates/*htmx"
+)
